move derive_randomenss to verification module
